moved set_warning_format to interface/utils/markdown_utiles.py and renamed to utils.py
