+++ aviary/interface/utils/utils.py
@@ -104,3 +104,37 @@ def write_markdown_variable_table(open_file, problem, outputs, metadata):
             units = '-'
         summary_line = f'| {var_name} | {val} | {units} |\n'
         open_file.write(summary_line)
+
+
+def set_warning_format(verbosity):
+    # if verbosity not set / not known yet, default to most simple warning format rather than no
+    # warnings at all
+    if verbosity is None:
+        verbosity = Verbosity.BRIEF
+
+    # Reset all warning filters
+    warnings.resetwarnings()
+
+    # NOTE identity comparison is preferred for Enum but here verbosity is often an int, so we need
+    # an equality comparison
+    if verbosity == Verbosity.QUIET:
+        # Suppress all warnings
+        warnings.filterwarnings('ignore')
+
+    elif verbosity == Verbosity.BRIEF:
+
+        def simplified_warning(message, category, filename, lineno, line=None):
+            return f'Warning: {message}\n\n'
+
+        warnings.formatwarning = simplified_warning
+
+    elif verbosity == Verbosity.VERBOSE:
+
+        def simplified_warning(message, category, filename, lineno, line=None):
+            return f'{category.__name__}: {message}\n\n'
+
+        warnings.formatwarning = simplified_warning
+
+    else:  # DEBUG
+        # use the default warning formatting
+        warnings.filterwarnings('default')
