+++ aviary/utils/functions.py
@@ -304,40 +304,6 @@ def create_printcomp(all_inputs: list, input_units: dict = {}, meta_data=_MetaDa
     return PrintComp
 
 
-def set_warning_format(verbosity):
-    # if verbosity not set / not known yet, default to most simple warning format rather than no
-    # warnings at all
-    if verbosity is None:
-        verbosity = Verbosity.BRIEF
-
-    # Reset all warning filters
-    warnings.resetwarnings()
-
-    # NOTE identity comparison is preferred for Enum but here verbosity is often an int, so we need
-    # an equality comparison
-    if verbosity == Verbosity.QUIET:
-        # Suppress all warnings
-        warnings.filterwarnings('ignore')
-
-    elif verbosity == Verbosity.BRIEF:
-
-        def simplified_warning(message, category, filename, lineno, line=None):
-            return f'Warning: {message}\n\n'
-
-        warnings.formatwarning = simplified_warning
-
-    elif verbosity == Verbosity.VERBOSE:
-
-        def simplified_warning(message, category, filename, lineno, line=None):
-            return f'{category.__name__}: {message}\n\n'
-
-        warnings.formatwarning = simplified_warning
-
-    else:  # DEBUG
-        # use the default warning formatting
-        warnings.filterwarnings('default')
-
-
 def promote_aircraft_and_mission_vars(group):
     """Promotes inputs and outputs in Aircraft and Mission hierarchy categories for provided group."""
     external_outputs = []
